Added Static Interface detail view
It lists the mac and vrf and what not.  I am a bit concerned that if all of the
fields were filled it it would be really crowded.  I am considering writting a
template for this case rather than just using the dhcp_details.html generic
template.  That way I can have details about the mac and ip and dns being
enabled in one table and then another regarding the vrf and workgroup.  It
might make sense to break up one table for dns related fun facts and another
for dhcp.  Dunno.